requirements_api.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pandas